tiny reorganize frontend files
