+++ frontend/components/color/PaletteColor.js
